+++ backend/game/logic/game.go
@@ -2,10 +2,8 @@ package logic
 
 import (
 	"errors"
-	"log"
 	"math/rand"
 	"sync"
-	"time"
 )
 
 type Game struct {
@@ -15,15 +13,12 @@ type Game struct {
 	trickColor        string
 	players           map[string]*Player
 	order             []string
-	starterTrun       int
-	distributorTurn   int
 	trickCallingCache *TrickCalling
-	turn              int
-	round             int
-	paused            bool
+	turner            turner
+	inactiveReset     chan struct{}
 }
 
-func NewGame(names []string) (*Game, error) {
+func NewGame(names []string, inactiveReset chan struct{}) (*Game, error) {
 	if len(names) < 3 {
 		return nil, errors.New("not enough players in the lobby")
 	} else if len(names) > 6 {
@@ -34,10 +29,12 @@ func NewGame(names []string) (*Game, error) {
 	})
 	game := &Game{
 		allCards: standardCardSet(),
-		round:    1,
 		players:  make(map[string]*Player),
 		order:    names,
 		deck:     make([]*card, 0, len(names)),
+		turner: turner{
+			round: 1,
+		},
 	}
 	for _, name := range names {
 		game.players[name] = &Player{
@@ -54,14 +51,12 @@ func (game *Game) PlayCard(name string, id int) error {
 	defer game.lock.Unlock()
 	if game.isGameOver() {
 		return errors.New("game is over")
-	} else if name != game.order[game.turn] {
+	} else if name != game.playerTurn() {
 		return errors.New("not your turn")
 	} else if game.trickCallingCache != nil {
 		return errors.New("hold on a minute")
-	} else if game.paused {
-		return errors.New("game is paused")
 	}
-	player := game.players[game.order[game.turn]]
+	player := game.players[name]
 	card, ok := player.hand[id]
 	if !ok {
 		return errors.New("you don't have that card")
@@ -70,87 +65,18 @@ func (game *Game) PlayCard(name string, id int) error {
 	if !card.isPlayable(getDeckColor(game.deck), player.hand) {
 		return errors.New("card is not playable; look harder")
 	}
+	if game.inactiveReset != nil {
+		game.inactiveReset <- struct{}{}
+	}
 	// play
 	delete(player.hand, id)
 	game.deck = append(game.deck, card)
-	game.turn = (game.turn + 1) % len(game.players)
-	// trick is finished
-	if game.starterTrun == game.turn {
-		game.paused = true
-		go func() {
-			time.Sleep(3 * time.Second)
-			game.lock.Lock()
-			defer game.lock.Unlock()
-			game.finishTrick()
-			// round is over
-			if len(player.hand) <= 0 {
-				game.finishRound()
-				if !game.isGameOver() {
-					game.prepareRound()
-				} else {
-					for _, player := range game.players {
-						log.Printf("player=%s has %d points", player.Name, player.Points)
-					}
-				}
-			}
-			game.paused = false
-		}()
-	}
+	game.advanceTurn()
 	return nil
 }
 
-func (game *Game) finishTrick() {
-	trickColor := getTrickColor(game.trickColor, game.deck)
-	deckColor := getDeckColor(game.deck)
-	log.Printf("game trick=%s round trick=%s", game.trickColor, trickColor)
-	best := game.deck[0]
-	bestIndex := 0
-	for i, card := range game.deck[1:] {
-		if card.calculateValue(trickColor, deckColor) > best.calculateValue(trickColor, deckColor) {
-			best = card
-			bestIndex = i + 1
-		}
-	}
-	bestIndex = (bestIndex + game.starterTrun) % len(game.players)
-	log.Printf("best was player at %d with %s value of %d", bestIndex, best.location, best.calculateValue(trickColor, deckColor))
-	player := game.players[game.order[bestIndex]]
-	player.Trick.Actual++
-	game.starterTrun = bestIndex
-	game.turn = bestIndex
-	game.deck = game.deck[0:0]
-}
-
-func (game *Game) finishRound() {
-	game.distributorTurn = (game.distributorTurn + 1) % len(game.players)
-	game.round++
-	for _, player := range game.players {
-		if player.Trick.Actual == player.Trick.Called {
-			player.Points += 10 + 20*player.Trick.Called
-		} else {
-			diff := player.Trick.Called - player.Trick.Actual
-			if diff < 0 {
-				diff *= -1
-			}
-			player.Points -= 10 + 20*diff
-		}
-	}
-}
-
-func (game *Game) prepareRound() {
-	for _, player := range game.players {
-		player.Trick = nil
-	}
-	game.distributeCards()
-	game.starterTrun = (game.distributorTurn + 1) % len(game.players)
-	game.turn = game.starterTrun
-	game.trickCallingCache = &TrickCalling{
-		PlayersLeft: len(game.players),
-		Round:       game.round,
-	}
-}
-
 func (game *Game) isGameOver() bool {
-	return game.round*len(game.players) > len(game.allCards)
+	return game.turner.round*len(game.players) > len(game.allCards)
 }
 
 func (game *Game) Deck() []string {
@@ -172,27 +98,21 @@ func (game *Game) DeckColor() *string {
 	return nil
 }
 
-func (lobby *Game) PlayerTurn() string {
-	lobby.lock.Lock()
-	defer lobby.lock.Unlock()
-	return lobby.order[lobby.turn]
-}
-
 func (game *Game) distributeCards() {
 	rand.Shuffle(len(game.allCards), func(i, j int) {
 		game.allCards[i], game.allCards[j] = game.allCards[j], game.allCards[i]
 	})
 	var i int
 	for _, player := range game.players {
-		for _, card := range game.allCards[i*game.round : (i+1)*game.round] {
+		for _, card := range game.allCards[i*game.turner.round : (i+1)*game.turner.round] {
 			player.hand[card.id] = card
 		}
 		i++
 	}
-	if len(game.players)*game.round == len(game.allCards) {
+	if len(game.players)*game.turner.round == len(game.allCards) {
 		i = rand.Intn(len(game.allCards))
 	} else {
-		i = len(game.players) * game.round
+		i = len(game.players) * game.turner.round
 	}
 	game.deck = game.deck[0:0]
 	game.trickColor = game.allCards[i].color