+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## [Unreleased]
+
 ## [v0.2.0] - 2020-02-03
 ### Added
 - Permanent notification when players turn
@@ -20,4 +22,5 @@
 - Frontend
 - Backend
 
+[Unreleased]: https://github.com/terrakuh/wizard/compare/v0.2.0...master
 [v0.2.0]: https://github.com/terrakuh/wizard/compare/v0.1.0...v0.2.0