+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [Unreleased]
+## [v0.3.0] - 2020-02-25
 ### Added
 - Lobby lifetime and timeouts
 - Notification settings
@@ -32,5 +32,6 @@
 - Frontend
 - Backend
 
-[Unreleased]: https://github.com/terrakuh/wizard/compare/v0.2.0...master
+[Unreleased]: https://github.com/terrakuh/wizard/compare/v0.3.0...dev
+[v0.3.0]: https://github.com/terrakuh/wizard/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/terrakuh/wizard/compare/v0.1.0...v0.2.0